+++ Sources/KeyboardKit/UIKit/Buttons/NextKeyboardUIButton.swift
@@ -12,7 +12,6 @@ import UIKit
  This button switches to the next system keyboard when it is
  tapped, and opens a list of system keyboards when pressed.
  */
-@available(iOS 13.0, *)
 public class NextKeyboardUIButton: UIButton {
     
     public convenience init(
@@ -28,6 +27,6 @@ public class NextKeyboardUIButton: UIButton {
             scale: scale)
         setImage(image, for: [])
         self.tintColor = tintColor
-        controller.setupNextKeyboardButton(self)
+        setupAsNextKeyboardButton(in: controller)
     }
 }